Added FieldExtensions
Added a new internal static class `FieldExtensions` with two methods: `GetCssClasses` and `Serialize`. `GetCssClasses` returns a string of CSS classes for a `Field` object, while `Serialize` returns a string representation of a `Field` object. Also, removed the code that was adding a new `Tuple<int, string>` to the `_currentMove` list in the `Playground` class.