`Add GetOrDefault extension method to DictionaryExtensions`
Introduces a new namespace `CodeBreaker.Blazor.Client.Extensions` and within this, a new static class `DictionaryExtensions` is created. This class includes a new generic extension method `GetOrDefault` for any types of `TKey` and `TValue`. This method takes a dictionary and a key as parameters, attempts to retrieve the value associated with the key from the dictionary, and returns the associated value if found, or the default value for `TValue` if not.