Refactor HttpClient and update configuration properties
This commit includes significant changes to the HttpClient configuration and the application's configuration properties. The HttpClient for "GameApi" has been replaced with two new HttpClients, `IGamerNameSuggestionClient` and `IGamesClient`, each configured with their respective base addresses "UserApiBase" and "GameApiBase". The "ApiBase" configuration property has been renamed to "GameApiBase" and a new "UserApiBase" property has been added, which is set to "https://localhost:7133/" in the development settings. The `AddHttpClient` calls have been relocated from `AddRazorComponents` to `AddMsalAuthentication` in `Program.cs`. No other changes have been made to the `AddMsalAuthentication` method, and the `IMobileDetectorService` continues to be added as a scoped service. The "DetailedErrors" configuration property remains set to true.